+++ Makefile
@@ -14,8 +14,7 @@ LDFLAGS = -extldflags \
 		  -X "main.BuildNumber=$(BUILDNUMER)"
 
 GOBUILD=CGO_ENABLED=0 go build -trimpath -ldflags "$(LDFLAGS)"
-CLI=cli.go
-SERVE=serve.go
+CLI=cmd/cli.go
 
 PLATFORM_LIST = \
 	linux-amd64 \
@@ -28,27 +27,21 @@ all: linux-amd64 linux-arm64 darwin-amd64 darwin-arm64 windows-amd64
 
 build:
 	go build -ldflags "$(LDFLAGS)" -o bin/norn $(CLI)
-	go build -ldflags "$(LDFLAGS)" -o bin/norns $(SERVE)
 
 linux-arm64:
-	GOARCH=arm64 GOOS=linux $(GOBUILD) -o $(BINDIR)/$(NAME)-$@-cli $(CLI)
-	GOARCH=arm64 GOOS=linux $(GOBUILD) -o $(BINDIR)/$(NAME)-$@-serve $(SERVE)
+	GOARCH=arm64 GOOS=linux $(GOBUILD) -o $(BINDIR)/$(NAME)-$@ $(CLI)
 
 linux-amd64:
-	GOARCH=amd64 GOOS=linux $(GOBUILD) -o $(BINDIR)/$(NAME)-$@-cli $(CLI)
-	GOARCH=amd64 GOOS=linux $(GOBUILD) -o $(BINDIR)/$(NAME)-$@-serve $(SERVE)
+	GOARCH=amd64 GOOS=linux $(GOBUILD) -o $(BINDIR)/$(NAME)-$@ $(CLI)
 
 darwin-amd64:
-	GOARCH=amd64 GOOS=darwin $(GOBUILD) -o $(BINDIR)/$(NAME)-$@-cli $(CLI)
-	GOARCH=amd64 GOOS=darwin $(GOBUILD) -o $(BINDIR)/$(NAME)-$@-serve $(SERVE)
+	GOARCH=amd64 GOOS=darwin $(GOBUILD) -o $(BINDIR)/$(NAME)-$@ $(CLI)
 
 darwin-arm64:
-	GOARCH=arm64 GOOS=darwin $(GOBUILD) -o $(BINDIR)/$(NAME)-$@-cli $(CLI)
-	GOARCH=arm64 GOOS=darwin $(GOBUILD) -o $(BINDIR)/$(NAME)-$@-serve $(SERVE)
+	GOARCH=arm64 GOOS=darwin $(GOBUILD) -o $(BINDIR)/$(NAME)-$@ $(CLI)
 
 windows-amd64:
-	GOARCH=amd64 GOOS=windows $(GOBUILD) -o $(BINDIR)/$(NAME)-$@-cli.exe $(CLI)
-	GOARCH=amd64 GOOS=windows $(GOBUILD) -o $(BINDIR)/$(NAME)-$@-serve.exe $(SERVE)
+	GOARCH=amd64 GOOS=windows $(GOBUILD) -o $(BINDIR)/$(NAME)-$@.exe $(CLI)
 
 lint:
 	GOOS=darwin golangci-lint run ./...
