+++ internal/common/helper.go
