--- serve.go
@@ -1,11 +1,0 @@
-package main
-
-import (
-	"context"
-	"github.com/kentio/norn/internal"
-)
-
-func main() {
-	app := internal.NewApp(context.Background())
-	app.Run()
-}
