--- service/api/fx.go
@@ -1,8 +1,0 @@
-package api
-
-import "go.uber.org/fx"
-
-var Model = fx.Options(
-	fx.Provide(NewRouter),
-	fx.Provide(NewApi),
-)
