+++ service/pick/errors.go
@@ -1,4 +1,4 @@
-package feature
+package pick
 
 import "errors"
