+++ service/pick/pick.go
@@ -1,4 +1,4 @@
-package feature
+package pick
 
 import (
 	"context"
